fix a bug

+++ cityflow_database/utils/processor.py
@@ -94,9 +94,10 @@ def delete_module(id):
     module = get_module(id)
     if module:
         icon = module.get('icon')
-        if icon.startswith('/api/dataset/source'):
-            icon_path = os.path.join(source_folder,'icons',os.path.basename(icon))
-            delete_file(icon_path)
+        if icon:
+            if icon.startswith('/api/dataset/source'):
+                icon_path = os.path.join(source_folder,'icons',os.path.basename(icon))
+                delete_file(icon_path)
         files = module.get('files',[])
         for file in files:
             if file['data'].startswith('/api/dataset/source'):
